(#1175) fix ubuntu16_04 check script
This fixes an issue I managed to overlook when I did the initial
add of the "ubuntu16_04" platform.  The platform-specific check
scripts need to exit cleanly, returning either 0 or 1 as appropriate
depending on whether the check script is a match for the current
platform.  The shell's immediate "exit-on-error" flag is not set for
the context in which the check script is executed, so we have to
explicitly call "exit 1" early if warranted.

[ci skip]