(#1174) add Ubuntu i386 architecture support
This commit captures a few critical steps that I performed almost
as an afterthought during the process of adding Ubuntu 16.04 LTS
as a development platform.  Support for searching and installing
from the i386 repositories has to be explicitly enabled by adding
i386 as a "foreign" architecture using "dpkg".  This is best done
in the "ubuntu16_04-packages.sh" script.

[ci skip]