#1042 initial commit for adding centos8 as an RCC platform
This includes the expected new "centos8" directory under
"projects/core/rcc/platforms", but also includes a fix for a subtle
error in "util.mk" that has evidently existed for some time but was
not discovered until it broke the "centos8" build.

The "makeExportLinks.sh" and "install-opencpi.sh" changes are (should
be) duplicates of changes that are pending in another branch waiting
to be merged with "develop".  They eliminate early dependencies on
"python" and "make" during bootstrapping of a fresh installation of
OpenCPI from source.

Finally, the change to "application.mk" is purely cosmetic, and fixes
an indentation annoyance.