(#1284) add doc RPM to existing packaging machinery
This commit is mostly about being able to run "make rpm Package=doc"
from the top-level OpenCPI directory and end up with a RPM more or
less identical to what can be created with the "releng" packaging
machinery.  "Makefile", "driver.spec", and "hw.spec" had a bit of
cosmetic cleanup inflicted on them.

Of particular note are the modifications made to "export-utils.sh".
The RPM packaging machinery hates broken symlinks, and we should
not be creating symlinks to things that do not exist.  Issue #1273
resolution had unintended consequences.

[ci skip]