(#1314) define YUM variable for CentOS 6 and 7 package installation
This commit primarily fixes a bug reported for the "centos7"
development platform.  The reporter was attempting to install
OpenCPI on a system without Internet connectivity (local yum
repositories), and the "swig3" package was unavailable.

The second aspect of this commit is replacing invocations of
"yum (several long options)" with "$YUM" where
YUM="yum --assumeyes --setopt=skip_missing_names_on_install=False".
It is safe to invoke "yum" this way regardless of the associated
command (list, install, erase).  Note an equivalent change was not
made to "centos8-packages.sh" because "dnf" is doing the right
thing by default.

[ci skip]