(#1284) slight refactoring to do all the building in "make"
Round two.  This commit addresses the legitimate concern about
splitting the building of objects for packaging between "make"
and "rpmbuild".  One or the other should handle the build task.

Also, "make" should not be called, indirectly from a spec file helper
script or otherwise, given that it's "make" calling "rpmbuild" in the
first place.

[ci skip]