(#1336) leave a HDL-platform-supplied "system.xml" alone
This commit fixes a subtle bug in the way "deploy-platform.sh"
populates the SD card directory.  If the RCC platform provides
a "system.xml" file, it overwrites the one provided by the HDL
platform (if any).  Even worse, it clobbers the original HDL-
platform-provided file due to the copy operation's destination
being a symlink.