(#1349) revert some of #1336 changes
This commit removes the logic implementing an ill-fated attempt to
have "setVarsFromMake" set and export OCPI_ALL_{HDL,RCC}_PLATFORMS.

Issue is "eval $vars" at the end of the "setVarsFromMake" function
and how things like "a=b c=d e=f export g=$a" get evaluated.  For
this example, the result of the eval is simply an exported variable
"g" set to "$a", i.e., none of the other variable assignments are
visible outside the context of the "export" command.  Bottom line:
cannot have any "export g=$a" statements in "{hdl,rcc}-targets.mk".

[ci skip]