(#1351) unset OCPI_ALL_{RCC,HDL}_PLATFORMS after setVarsFromMake call
This commit restores platform installation functionality that breaks
if the OCPI_ALL_{RCC,HDL}_PLATFORMS variables are set in some contexts.
Essentially, setting and exporting those variables in makefiles is
temporarily a no-op because their being set is an indication that
platform installation has completed successfully, and the database
won't be updated further.

[ci skip]