(#1454) add new OCPI_DISTRO_BUILD environment variable
This is the initial commit implementing a new OCPI_DISTRO_BUILD
environment variable.  When set to "1", it is an indicator to
the build system that "lowest common denominator" build flags
for the x86_64 architecture (appropriate for binary distribution
packages such as RPMs, tarballs, etc.) should be used.

As of this writing, the only thing affected by OCPI_DISTRO_BUILD
is the GNU MP (gmp) build.

The environment variable can be set and exported prior to running
"install-opencpi.sh" or "ocpiadmin".  Alternatively, specifying
"-d" or "--distro" as an argument to either script will set and
export the OCPI_DISTRO_BUILD variable.

The mods to "namespace.txt" and the core RCC platform "*.mk" files
are to correct minor typos that had previously escaped notice.