(#1360) initial commit of "socket_write.rcc" changes
This commit is mostly about eliminating the "socket_write" RCC worker's
dependency on Boost.  This results in a clean reorganization of the
worker source tree, not to mention a significant simplification of the
worker's "Makefile".  Most of the file additions/deletions have to do
with relocating the "asio" headers, which, unfortunately, we have to
continue using for the time being.  For supported RCC development
platforms, we can get away with using a distro-supplied "asio-devel"
package, but that doesn't work easily for cross-builds.

Unrelated changes "buried" in this commit have to do with how the
Xilinx kernel source repos get cloned/downloaded.  The key change
here is an updated error message in "createLinuxKernelHeaders.sh"
that no longer hints at the presence of code under development
that may never see the light of day.

[ci skip]