(#2484) more python3.5 portability fixes
This commit includes the fix for issue #2483 as well.  Python3.5
is pickier than python3.6+ about allowing the use of path objects
as arguments to "os.*()" methods: according to TFM, the arguments
are supposed to be string objects regardless of the python3 minor
version.  Also, f-strings weren't introduced until python3.6. The
need to worry about python3.5 portability will go away when we can
drop support for `ubuntu16_04`.

[ci skip]