(#2702) initial mods for Tutorial 9
This commit is primarily about renaming the "counter" worker
to "tcounter" to avoid a naming collision with a HDL primitive.
Also, "orig" and "debug" versions of the various worker config
files have been created to allow tutorial users to restore the
original state of the worker.  The tutorial doc itself is a
work in progress.

[ci skip]